+++ components/src/main/java/org/patternfly/component/drawer/DrawerSection.java
@@ -16,6 +16,7 @@
 package org.patternfly.component.drawer;
 
 import org.jboss.elemento.ElementTextMethods;
+import org.patternfly.style.Modifiers.Plain;
 
 import elemental2.dom.HTMLDivElement;
 
@@ -28,7 +29,8 @@ import static org.patternfly.style.Classes.section;
  * An optional content above the {@link DrawerContent} and {@link DrawerPanel} in a {@link Drawer} component.
  */
 public class DrawerSection extends DrawerSubComponent<HTMLDivElement, DrawerSection> implements
-        ElementTextMethods<HTMLDivElement, DrawerSection> {
+        ElementTextMethods<HTMLDivElement, DrawerSection>,
+        Plain<HTMLDivElement, DrawerSection> {
 
     // ------------------------------------------------------ factory