+++ components/src/main/java/org/patternfly/component/form/Checkbox.java
@@ -27,6 +27,7 @@ import org.patternfly.component.BaseComponent;
 import org.patternfly.component.ComponentType;
 import org.patternfly.component.HasValue;
 import org.patternfly.handler.ChangeHandler;
+import org.patternfly.core.Aria;
 import org.patternfly.style.Classes;
 import org.patternfly.style.Modifiers.Disabled;
 import org.patternfly.style.Modifiers.Required;
@@ -139,6 +140,12 @@ public class Checkbox extends BaseComponent<HTMLElement, Checkbox> implements
                 labelElement.classList.remove(modifier(Classes.disabled));
             }
         }
+        return this;
+    }
+
+    /** Sets the {@code aria-describedby} attribute on the input element. */
+    public Checkbox ariaDescribedBy(String describedBy) {
+        inputElement.setAttribute(Aria.describedBy, describedBy);
         return this;
     }